+++ src/utils/throttle.js
@@ -1,0 +1,13 @@
+const throttle = (func, ms) => {
+  let throttled = false;
+  return (...args) => {
+    if (!throttled) {
+      throttled = true;
+      setTimeout(() => {
+        func(...args);
+        throttled = false;
+      }, ms);
+    }
+  };
+};
+export default throttle;
